Patch 5: support setting NSID option (SNA-20345)
To ease debugging behind anycast, add support to dnsdist for
returning an identifier of the POP that is handling the request.

This commit introduces a new ``SetEDNSOptionResponseAction``
action that works similarly to ``SetEDNSOptionAction``, but that
can be used for responses.

Example:
addResponseAction(
    EDNSOptionRule(EDNSOptionCode.NSID),
    SetEDNSOptionResponseAction(EDNSOptionCode.NSID, "foobar")
)

In the above, the NSID option will be set when the question has
this bit set. Note that it will override what is sent by the DNS
server.