Try harder matching command names in pdnsutil.
We would like to eventually move towards a consistent naming of the pdnsutil
commands, in the form:
  pdnsutil <object (a noun)> <action (a verb)> <arguments>
with no dash between the object and the action.

However the existing commands are currently in the form action-object.
This change attempts to recognize both "object-action" and "object action"
and rewrites them as "action-object" to match the current command set, as
a first step.