dnsdist: YAML: Correct `enable_proxy_protocol` value for TCP/UDP binds
The YAML `enable_proxy_protocol` property of `BindConfiguration` is
documented as defaulting to "true", but is actually defined to default
to false. This commit redefines this property to default to "true",
which matches the documentation. It also matches the documentation of
the `enableProxyProtocol` option to the Lua `addLocal()` function.

However, when `loadBinds()` constructs the `ClientState` for a TCP
or UDP bind, it hardcodes the last parameter (`enableProxyProtocol`)
to false, i.e. it ignores the configured value. In this commit, the
`enableProxyProtocol` parameter to the `ClientState` constructor is
taken from the corresponding value loaded by the YAML parser.