Overhaul pdnsutil command name parsing again.
We are moving towards a more consistent command naming scheme, where the
general syntax is:
  pdnsutil object-type command args...
where 'object-type' is a noun and 'command' is a verb.

For example, in this new world order, "pdnsutil list-zone" becomes
"pdnsutil zone list", "pdnsutil set-meta" becomes "pdnsutil metadata
set", etc.

The old world order commands are still recognized and their behaviour is
not modified.

Signed-off-by: Miod Vallat <miod.vallat@powerdns.com>