rec: allow negation in outgoing.edns_subnet_allow_list
Casting input part to Netmask does not allow to insert negation in
s_ednsremotesubnets list as ComboAddress would not allow, however if we
directly use addMask from NetmaskGroup negation are handled properly. In
case the input is not a valid netmask, catch() will still add the string
as a domain.