rec: only store IP (and no port) as local address in cookie store
Otherwise binding a local TCP port will fail, as the stored port
is likely still in use do to connections lingering.

Observed when forcing DoT to specific nameservers that support
cookies.  In many cases the effect was: see the bind call failing and
go to the next nameserver, that's likely why it was noticed before.

Signed-off-by: Otto Moerbeek <otto.moerbeek@open-xchange.com>