Structured logging support for the Authoritative server.
Setup a syslog structured logger for pdns-server and pdnsutil,
copied from recursor.

Add structured logging flavour of existing log messages in the
Authoritative server, except in tools (ixfrdist and pdnsutil).

Structured logging is not enabled yet in the server or any of its tools
and can not be enabled yet.

Signed-off-by: Miod Vallat <miod.vallat@powerdns.com>