dnsdist: Better performance when using `recvmmsg`
`recvmmsg` and `sendmmsg` are Linux-specific syscalls to receive
and send multiple messages on a socket via a single call, which
in theory improves performance a lot, especially with the
speculative execution vulnerabilities mitigations making syscalls
expensive.
We had several reports that turning this feature on in DNSdist did
not really improve performance, which is quite unexpected. Looking
at a CPU profile in `perf`, we were spending an awful lot of time
resizing buffers that did not need to be resized most of the time
because they had not been used. This commit ensures that we only
resize and reset buffers for slots that have been used in the
previous round.

With that I'm getting a nice performance boost from using `recvmmsg`
and `sendmmsg` while processing 80k QPS at 100% cache-hit ratio
from a single thread: the CPU usage shrinks by roughly 15%.

Signed-off-by: Remi Gacogne <remi.gacogne@powerdns.com>
(cherry picked from commit c9e3cb9e619b16e50021116272e783bac533cc12)