dnsdist: Raise the maximum number of descriptors to 1M
When running under systemd there is a cap on the number of open file or
socket descriptors that we can have at a given time. We used to set this
to 16k which was a large enough value for most installations, but now that
people are deploying DNSdist to offer DoT/DoH/DoQ/DoH3 to large numbers of
users we are regularly getting complaints that the default value is too low.
On the other hand I'm not aware of any case where having a cap actually
prevented an issue where we would be opening too many descriptors.
Therefore this commit is raising the default value to 1 millions, a large
enough values for even very large setups dealing with > 100k incoming
connections at a time.

Signed-off-by: Remi Gacogne <remi.gacogne@powerdns.com>
(cherry picked from commit 91f86e3950f9a59a6da6d2937e84d8a3186f2789)