m4/pdns_check_libcrypto: fix ecdsa/eddsa includedir
On more excotic Linux distributions, like NixOS, the directory in which
the libraries are stored and the directory of the headers is different.
This uses pkg-config to find the right includedir of the headers.

Signed-off-by: famfo <famfo@famfo.xyz>
(cherry picked from commit 48578f782de8d6c840fa41b8e1eed2a12ed7c1c4)