dnsdist: Build our Rust lib in `dev` profile when `CARGO_USE_DEV` is set
It is roughly halving the compile time in my tests, and comes
with more checks (see https://doc.rust-lang.org/cargo/reference/profiles.html#dev)
so I find it useful to be able to explicitly request building
using the `dev` profile. The default remains to build in `release`
mode for performance, of course.

Signed-off-by: Remi Gacogne <remi.gacogne@powerdns.com>