Make pdnsutil zone list abort if output is stalled for 30 seconds.
We have noticed that some container solutions do not appear to implement
proper pipe semantics, and it is possible to end up with pdnsutil processes
stuck sleeping because the pipe no longer has any readers, but no SIGPIPE
is generated.

In order to mitigate this, we switch stdout to non-blocking if output is not
a tty, and monitor output progress; if output stalls for an arbitrary 30
seconds, we abort the listing operation. This allows us to not lead an open
read transaction on LMDB backends.

Signed-off-by: Miod Vallat <miod.vallat@powerdns.com>