dnsdist: Fix handling of IP-only TLS certificates
To properly handle SNI with OpenSSL, we have to keep track of the TLS
names supported by the different certificates associated to a frontend.
Since IP-only names are not involved in SNI negotiations, we are currently
skipping them when creating the TLS SNI map. Unfortunately this causes
an issue when we have a single certificate that only contains IPs,
the default context not being properly set.
This commit ensures the default context is properly set in that case.

Signed-off-by: Remi Gacogne <remi.gacogne@powerdns.com>
(cherry picked from commit 99faabf75f11a0eec0720327e0753cbfc31b9d22)